refactor(bin): move Pep/10 CPU stub to separate file
Trying to avoid a lava-flow pattern here.
