+++ bin/simexp/main.cpp
@@ -1,6 +1,7 @@
 
 #include <cstdio>
 #include <string>
+#include "./pep10isa.hpp"
 #include "./simuloop.hpp"
 #include "core/ds/hash/djb.hpp"
 #include "core/integers.h"
